+++ src/ScottPlot5/ScottPlot5 Controls/ScottPlot.Avalonia/AvaPlot.cs
@@ -5,17 +5,20 @@ using Avalonia.Media;
 using Avalonia.Platform;
 using Avalonia.Rendering.SceneGraph;
 using Avalonia.Threading;
-using ScottPlot.Control;
 using SkiaSharp;
 
 using Controls = Avalonia.Controls;
+using System;
 
 namespace ScottPlot.Avalonia;
+
+#pragma warning disable CS0618 // disable obsolete warnings
 
 public class AvaPlot : Controls.Control, IPlotControl
 {
     public Plot Plot { get; internal set; }
 
+    [Obsolete("Deprecated. Use UserInputProcessor instead. See ScottPlot.NET demo and FAQ for usage details.")]
     public IPlotInteraction Interaction { get; set; }
     public IPlotMenu? Menu { get; set; }
     public Interactivity.UserInputProcessor UserInputProcessor { get; }
@@ -29,7 +32,7 @@ public class AvaPlot : Controls.Control, IPlotControl
         Plot = new() { PlotControl = this };
         ClipToBounds = true;
         DisplayScale = DetectDisplayScale();
-        Interaction = new Interaction(this);
+        Interaction = new Control.Interaction(this); // TODO: remove in an upcoming release
         UserInputProcessor = new(Plot);
         Menu = new AvaPlotMenu(this);
         Focusable = true; // Required for keyboard events