+++ src/ScottPlot5/ScottPlot5 Controls/ScottPlot.WPF/WpfPlotBase.cs
@@ -1,10 +1,12 @@
-﻿using System.Windows;
+﻿using System;
+using System.Windows;
 using System.Windows.Input;
 using System.Windows.Media;
-using ScottPlot.Control;
 using SkiaSharp;
 
 namespace ScottPlot.WPF;
+
+#pragma warning disable CS0618 
 
 public abstract class WpfPlotBase : System.Windows.Controls.Control, IPlotControl
 {
@@ -12,6 +14,8 @@ public abstract class WpfPlotBase : System.Windows.Controls.Control, IPlotContro
     public abstract void Refresh();
 
     public Plot Plot { get; internal set; }
+
+    [Obsolete("Deprecated. Use UserInputProcessor instead. See ScottPlot.NET demo and FAQ for usage details.")]
     public IPlotInteraction Interaction { get; set; }
     public float DisplayScale { get; set; }
     public IPlotMenu? Menu { get; set; }
@@ -29,7 +33,7 @@ public abstract class WpfPlotBase : System.Windows.Controls.Control, IPlotContro
     {
         Plot = new Plot() { PlotControl = this };
         DisplayScale = DetectDisplayScale();
-        Interaction = new Interaction(this);
+        Interaction = new Control.Interaction(this); // TODO: remove in an upcoming release
         UserInputProcessor = new(Plot);
         Menu = new WpfPlotMenu(this);
         Focusable = true;