+++ src/ScottPlot5/ScottPlot5 Controls/ScottPlot.WinForms/FormsPlotBase.cs
@@ -1,5 +1,4 @@
-﻿using ScottPlot.Control;
-using ScottPlot.Interactivity;
+﻿using ScottPlot.Interactivity;
 using SkiaSharp;
 using System;
 using System.Diagnostics;
@@ -8,12 +7,15 @@ using System.Windows.Forms;
 
 namespace ScottPlot.WinForms;
 
+#pragma warning disable CS0618 
+
 public abstract class FormsPlotBase : UserControl, IPlotControl
 {
     public abstract GRContext GRContext { get; }
 
     public Plot Plot { get; internal set; }
 
+    [Obsolete("Deprecated. Use UserInputProcessor instead. See ScottPlot.NET demo and FAQ for usage details.")]
     public IPlotInteraction Interaction { get; set; }
     public IPlotMenu? Menu { get; set; }
     public UserInputProcessor UserInputProcessor { get; }
@@ -24,7 +26,7 @@ public abstract class FormsPlotBase : UserControl, IPlotControl
     {
         Plot = new() { PlotControl = this };
         DisplayScale = DetectDisplayScale();
-        Interaction = new Interaction(this);
+        Interaction = new Control.Interaction(this); // TODO: remove in an upcoming release
         UserInputProcessor = new(Plot);
         Menu = new FormsPlotMenu(this);